rearrangement and wired sender to checker
