+++ utils/blogpost_checker.py
@@ -3,6 +3,10 @@ from bs4 import BeautifulSoup
 from datetime import datetime
 from urllib.request import urlopen
 
+from utils.update_status import send_tweet
+from utils.config import Config
+from utils.arg_parser import initialize_parser
+
 feed_url = "https://avogadbro.blogspot.com/feeds/posts/default"
 
 
@@ -14,23 +18,24 @@ def get_post(blog_table: boto3.resource, post_id: str):
     return response.get("Item")
 
 
-def parse_feed():
+def parse_feed(cfg: Config):
     """
     Grabs RSS feed for the blog and looks for new entries.
     """
     with urlopen(feed_url) as data:  # nosec
         raw_xml = data.read()
 
-    publised_fmt = "%Y-%m-%dT%H:%M:%S.%f%z"
+    published_fmt = "%Y-%m-%dT%H:%M:%S.%f%z"
 
     parsed_xml = BeautifulSoup(raw_xml, "xml")
 
     # I'd be shocked if blogger didn't keep this in order but
     latest_published_at = 0
+    latest_post = None
     for post in parsed_xml.findAll("entry"):
         published_at = int(
             datetime.strptime(
-                post.select("published")[0].text, publised_fmt
+                post.select("published")[0].text, published_fmt
             ).timestamp()
         )
         if published_at > latest_published_at:
@@ -39,8 +44,12 @@ def parse_feed():
 
     dynamodb = boto3.resource("dynamodb")
     blog_table = dynamodb.Table("avobot_blog_posts")
+    if not latest_post:
+        # This really shouldn't happen but for completeness
+        print("No latest post. Maybe the RSS didn't load correctly?")
+
     latest_id = latest_post.select("id")[0].text
-    if not get_post(blog_table, latest_id):
+    if not get_post(blog_table, latest_id) or True:
         print("New blogpost detected; updating table.")
         post_title = latest_post.select("title")[0].text
         blog_table.put_item(
@@ -50,10 +59,17 @@ def parse_feed():
                 "title": post_title,
             }
         )
-    # tweet about it
+
+        send_tweet(
+            "@dvfeinblum just published a new blogpost! Check it out at https://avogadbro.blogspot.com/",
+            cfg,
+        )
+
     else:
         print("No new posts detected.")
 
 
+# To be removed at some point
 if __name__ == "__main__":
-    parse_feed()
+    args = initialize_parser()
+    parse_feed(Config(args.secrets))
